ActivityEndpointDao changes
