+++ schedule/schedule-appengine/src/main/java/com/jasify/schedule/appengine/dao/common/ActivityPackageActivityDao.java
@@ -1,7 +1,6 @@
 package com.jasify.schedule.appengine.dao.common;
 
 import com.google.appengine.api.datastore.Key;
-import com.google.appengine.api.datastore.Query;
 import com.jasify.schedule.appengine.dao.BaseCachingDao;
 import com.jasify.schedule.appengine.dao.BaseDaoQuery;
 import com.jasify.schedule.appengine.meta.activity.ActivityPackageActivityMeta;
@@ -9,7 +8,6 @@ import com.jasify.schedule.appengine.model.EntityNotFoundException;
 import com.jasify.schedule.appengine.model.activity.Activity;
 import com.jasify.schedule.appengine.model.activity.ActivityPackageActivity;
 import com.jasify.schedule.appengine.model.activity.ActivityType;
-import org.slim3.datastore.CompositeCriterion;
 import org.slim3.datastore.Datastore;
 
 import java.io.Serializable;
@@ -22,28 +20,36 @@ import java.util.List;
  */
 public class ActivityPackageActivityDao extends BaseCachingDao<ActivityPackageActivity> {
     private final ActivityTypeDao activityTypeDao = new ActivityTypeDao();
+    private final ActivityDao activityDao = new ActivityDao();
 
     public ActivityPackageActivityDao() {
         super(ActivityPackageActivityMeta.get());
     }
 
-    public List<ActivityPackageActivity> getBy(final Activity activity) {
+    public List<ActivityPackageActivity> getByActivityId(final Key activityId) throws EntityNotFoundException {
         ActivityPackageActivityMeta meta = getMeta();
-        try {
-            ActivityType activityType = activityTypeDao.get(activity.getActivityTypeRef().getKey());
-            final Key organisationKey = activityType.getOrganizationRef().getKey();
-            return query(new BaseDaoQuery<ActivityPackageActivity, ActivityPackageActivityMeta>(meta, new Serializable[0]) {
-                @Override
-                public List<Key> execute() {
-                    return Datastore.query(meta, organisationKey)
-                            .filter(new CompositeCriterion(meta,
-                                    Query.CompositeFilterOperator.AND,
-                                    meta.activityRef.equal(activity.getId()))).asKeyList();
-                }
-            });
-        } catch (EntityNotFoundException e) {
-            // Is this possible?
+
+        Activity activity = activityDao.get(activityId);
+        if (activity.getActivityTypeRef().getKey() == null) {
             return Collections.emptyList();
         }
+
+        Key activityTypeId = activity.getActivityTypeRef().getKey();
+        ActivityType activityType = activityTypeDao.get(activityTypeId);
+        if (activityType.getOrganizationRef().getKey() == null) {
+            return Collections.emptyList();
+        }
+
+        final Key organizationId = activityType.getOrganizationRef().getKey();
+
+        return query(new BaseDaoQuery<ActivityPackageActivity, ActivityPackageActivityMeta>(meta, new Serializable[0]) {
+            @Override
+            public List<Key> execute() {
+                return Datastore.query(
+                        Datastore.getCurrentTransaction(), meta, organizationId)
+                        .filter(meta.activityRef.equal(activityId))
+                        .asKeyList();
+            }
+        });
     }
 }
